Basic auth

+++ media-vault-app.API/Controllers/MediaEntriesController.cs
@@ -1,81 +1,103 @@
+using System.Security.Claims;
 using media_vault_app.Application.DTOs.MediaEntry.Request;
 using media_vault_app.Application.DTOs.MediaEntry.Response;
 using media_vault_app.Application.Interfaces.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace media_vault_app.API.Controllers
 {
     [ApiController]
+    [Authorize]
     [Route("[controller]")]
     public class MediaEntriesController : ControllerBase
     {
         private readonly IMediaEntryReadService _readService;
         private readonly IMediaEntryWriteService _writeService;
 
-        // Hardcoded user ID for testing (no auth yet)
-        //private static readonly Guid TestUserId = Guid.Parse("00000000-0000-0000-0000-000000000001");
-
-        public MediaEntriesController(IMediaEntryReadService readService, IMediaEntryWriteService writeService)
+        public MediaEntriesController(
+            IMediaEntryReadService readService,
+            IMediaEntryWriteService writeService)
         {
             _readService = readService;
             _writeService = writeService;
         }
 
-        [HttpPost("{userId:Guid}")]
+        [HttpPost]
         public async Task<ActionResult<MediaEntryDetailedDto>> CreateMediaEntry(
-            [FromRoute] Guid userId,
             [FromBody] MediaEntryCreateDto createDto,
             CancellationToken ct)
         {
+            if (!TryGetCurrentUserId(out var userId))
+                return Unauthorized();
+
             var result = await _writeService.CreateAsync(userId, createDto, ct);
 
-            return this.ToCreated(result, nameof(GetMediaEntryById), value => new { userId = value.UserId, id = value.Id });
+            return this.ToCreated(result, nameof(GetMediaEntryById), value => new { id = value.Id });
         }
 
-        [HttpGet("{userId:Guid}/{id:Guid}")]
+        [HttpGet("{id:Guid}")]
         public async Task<ActionResult<MediaEntryDetailedDto>> GetMediaEntryById(
-            [FromRoute] Guid userId,
             [FromRoute] Guid id,
             CancellationToken ct)
         {
+            if (!TryGetCurrentUserId(out var userId))
+                return Unauthorized();
+
             var result = await _readService.GetByIdAsync(userId, id, ct);
 
             return this.ToOk(result);
         }
 
-        [HttpGet("{userId:Guid}")]
+        [HttpGet]
         public async Task<ActionResult<IEnumerable<MediaEntryDetailedDto>>> GetMediaEntries(
-            [FromRoute] Guid userId,
             [FromQuery] int pageNumber = 1,
             [FromQuery] int pageSize = 10,
             CancellationToken ct = default)
         {
+            if (!TryGetCurrentUserId(out var userId))
+                return Unauthorized();
+
             var result = await _readService.GetDetailedCollectionAsync(userId, pageNumber, pageSize, ct);
 
             return this.ToOk(result);
         }
 
-        [HttpPut("{userId:Guid}/{id:Guid}")]
+        [HttpPut("{id:Guid}")]
         public async Task<IActionResult> UpdateMediaEntry(
-            [FromRoute] Guid userId,
             [FromRoute] Guid id,
             [FromBody] MediaEntryUpdateDto updateDto,
             CancellationToken ct)
         {
+            if (!TryGetCurrentUserId(out var userId))
+                return Unauthorized();
+
             var result = await _writeService.UpdateAsync(userId, id, updateDto, ct);
 
             return this.ToNoContent(result);
         }
 
-        [HttpDelete("{userId:Guid}/{id:Guid}")]
+        [HttpDelete("{id:Guid}")]
         public async Task<IActionResult> DeleteMediaEntry(
-            [FromRoute] Guid userId,
             [FromRoute] Guid id,
             CancellationToken ct)
         {
+            if (!TryGetCurrentUserId(out var userId))
+                return Unauthorized();
+
             var result = await _writeService.DeleteAsync(userId, id, ct);
 
             return this.ToNoContent(result);
+        }
+
+        private bool TryGetCurrentUserId(out Guid userId)
+        {
+            var userIdClaim = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (!Guid.TryParse(userIdClaim, out userId))
+            {
+                return false;
+            }
+            return true;
         }
     }
 }
